- Input Validation in Stickynote - No Multiple calls on server for same event - NotificationMessages in StickyNotes for all Events

+++ client/src/app/core/retrospective/retrospective-topic-board/sticky-note/sticky-note.component.ts
@@ -1,6 +1,8 @@
 import {Component, OnInit, Input} from '@angular/core';
 import {IconButtonType, NotifierService} from '../../../../shared';
 import {TopicService, IStickyNote, StickyNoteMode} from '../services/';
+import {NotificationMessage} from '../../../../shared/notification-message/notification-message';
+import {NotificationMessageType} from '../../../../shared/notification-message/notification-message-type';
 
 @Component({
   selector: 'rsb-sticky-note',
@@ -20,7 +22,12 @@ export class StickyNoteComponent implements OnInit {
   private covered: boolean = false;
 
   private _isWaitingForCommit = false;
+  private _isWaitingForDelet = false;
+  private _isWatingForReload = false;
 
+  private validationErrorMessage: NotificationMessage;
+  private titleError: boolean = false;
+  private descError: boolean = false;
 
   constructor(private topicService: TopicService,
               private notificationService: NotifierService) {
@@ -31,20 +38,37 @@ export class StickyNoteComponent implements OnInit {
   }
 
   public saveStickyNote(): void {
-    if (!this._isWaitingForCommit) {
+    if (!this._isWaitingForCommit && this.inputValidation()) {
       this._isWaitingForCommit = true;
       this.topicService.saveComment(this.stickyNote).first().subscribe((notificationMessage) => {
         this._isWaitingForCommit = false;
         this.notificationService.pushNextMessage(notificationMessage);
+      }, e => {
+        this._isWaitingForCommit = false;
+        this.notificationService.pushNextMessage(new NotificationMessage(NotificationMessageType.ERROR,
+          'There was a error while trying to save comment', 10));
       });
     }
   }
 
   public aboardEdit(): void {
     if (this.stickyNote.uuid == null) {
-      this.topicService.deleteComment(this.stickyNote);
-    } else {
-      this.topicService.reloadStickyNote(this.stickyNote);
+      this.deleteComment();
+    } else if (!this._isWatingForReload) {
+      this._isWatingForReload = true;
+      this.topicService.reloadStickyNote(this.stickyNote)
+        .first()
+        .subscribe((notificationMessage: NotificationMessage) => {
+          this._isWatingForReload = false;
+          this.notificationService.pushNextMessage(notificationMessage);
+        }, e => {
+          this._isWatingForReload = false;
+          this.notificationService.pushNextMessage(new NotificationMessage(NotificationMessageType.ERROR,
+            'There was a error while trying to reload comment', 10));
+        });
+    }
+    if (this.validationErrorMessage != null) {
+      this.validationErrorMessage.setMessageExpired();
     }
   }
 
@@ -53,11 +77,38 @@ export class StickyNoteComponent implements OnInit {
   }
 
   public deleteComment() {
-    this.topicService.deleteComment(this.stickyNote);
+    if (!this._isWaitingForDelet) {
+      this._isWaitingForDelet = true;
+      this.topicService.deleteComment(this.stickyNote)
+        .first()
+        .subscribe((notificationMessage: NotificationMessage) => {
+          this.notificationService.pushNextMessage(notificationMessage);
+          this._isWaitingForDelet = false;
+        }, e => {
+          this.notificationService.pushNextMessage(new NotificationMessage(NotificationMessageType.ERROR,
+            'There was a error while trying to delete comment', 10));
+          this._isWaitingForDelet = false;
+        });
+    }
   }
 
   public vote(): void {
 
+  }
+
+  public inputValidation(): boolean {
+    this.titleError = this.stickyNote.title == null || this.stickyNote.title.trim() === '';
+    this.descError = this.stickyNote.description == null || this.stickyNote.description.trim() === '';
+
+    if (this.titleError || this.descError) {
+      this.validationErrorMessage = new NotificationMessage(NotificationMessageType.WARNING,
+        'There are missing infomration for this sticky-note');
+      this.notificationService.pushNextMessage(this.validationErrorMessage);
+      return false;
+    } else if (this.validationErrorMessage != null) {
+      this.validationErrorMessage.setMessageExpired();
+    }
+    return true;
   }
 
   public get isEditMode(): boolean {
